+++ FeyonaBank/Views/Shared/_Layout.cshtml
@@ -31,7 +31,7 @@
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
                     <li><a asp-area="" asp-controller="Transactions" asp-action="DepositAndWithdraw">Deposit and Withdraw</a></li>
-                    <li><a asp-area="" asp-controller="Transactions" asp-action="Transfer">Transfer</a></li>
+                    <li><a asp-area="" asp-controller="Transfer" asp-action="Transfer">Transfer</a></li>
                     @*<li><a asp-area="" asp-controller="Home" asp-action="About">About</a></li>
                         <li><a asp-area="" asp-controller="Home" asp-action="Contact">Contact</a></li>*@
                 </ul>
