+++ FeyonaBank/Views/Transfer/Transfer.cshtml
@@ -16,7 +16,7 @@
         <br />
 
         <div>
-            <form asp-controller="Transactions" asp-action="Transfer" method="post"
+            <form asp-controller="Transfer" asp-action="Transfer" method="post"
                   data-ajax="true"
                   data-ajax-method="POST"
                   data-ajax-mode="append"
